新增 users 相關；調整 posts get, post；全部刪除判斷

+++ controllers/posts.js
@@ -1,14 +1,45 @@
 const successHandler = require('../utility/successHandler');
 const errorHandler   = require('../utility/errorHandler');
 const PostModel      = require('../models/posts');
+const UserModel      = require('../models/users');
 const resMsg = require('../utility/responseMessage');
 
 const posts = {
   async getPosts(request, response){
-    const posts = await PostModel.getDB();
+    const query    = request.query;
+    const timeSort = query.timeSort===undefined ? '-createdAt' : 'createdAt';
+    const keyword  = query.q===undefined ? {} : { content :new RegExp(query.q) };
+    const posts = await PostModel.getAllDB(keyword, timeSort)
+      .catch(error => errorHandler(response, 400, { message: resMsg.somethingWrong, error }));
     successHandler(response, {data: posts});
   },
 
+  async getPost(request, response){
+    const id = request.params.postId;
+    try{
+      await PostModel.getOneDB(id)
+        .then(result => {
+          !result
+            ? errorHandler(response, 400, {
+              message: resMsg.noItem,
+              error
+            })
+            : successHandler(response, { data: result });
+        })
+        .catch(error => {
+          errorHandler(response, 400, {
+            message: resMsg.noItem,
+            error
+          });
+        });
+    }
+    catch(error){
+      errorHandler(response, 400, {
+        message: resMsg.noItem,
+        error
+      })
+    }
+  },
 
   async postPost(request, response){
     try {
@@ -21,18 +52,31 @@ const posts = {
         errorHandler(response, 400, {message: resMsg.contentRequired});
         return;
       }
+      if(data.image) {
+        if( !data.image.toString().match(/(\S+?(?:jpe?g|png|gif|webp))$/) )
+         errorHandler(response, 400, {message: resMsg.imageNotMatch}); 
+      }
+
+      const user = await UserModel.getOneDB(data.user)
+        .catch(error => errorHandler(response, 400, { message: resMsg.noUser, error}));
+      if(!user){
+        return errorHandler(response, 400, { message: resMsg.noUser});
+      }
+
       data.name    = data.name.trim();
       data.content = data.content.trim();
       PostModel.postDB(data)
         .then(async(result) => 
           successHandler(response, {
             message: resMsg.postSuccess, 
-            data: await PostModel.getDB()
+            data: await PostModel.getAllDB()
           }))
-        .catch(error => errorHandler(response, 400, {
-          message: resMsg.postFail,
-          error
-        }));
+        .catch(error => {
+          errorHandler(response, 400, {
+            message: resMsg.postFail,
+            error
+          })
+        });
     }
     catch(error) {
       errorHandler(response, 400, { 
@@ -68,7 +112,7 @@ const posts = {
               message: resMsg.noItem,
             }); 
           }
-          const posts = await PostModel.getDB()
+          const posts = await PostModel.getAllDB()
           successHandler(response, {
             message: resMsg.patchSuccess,
             data: posts, 
@@ -101,7 +145,7 @@ const posts = {
               message:resMsg.noItem,
             });   
           }
-          const posts = await PostModel.getDB()
+          const posts = await PostModel.getAllDB()
           successHandler(response, {
             message: resMsg.deleteSuccess,
             data: posts, 
@@ -124,8 +168,12 @@ const posts = {
   },
 
 
-
   async deletePosts(request, response){
+    if(request.originalUrl === '/posts/'){
+      return errorHandler(response, 400, {
+        message: resMsg.noItem, 
+      });
+    }
     PostModel.deleteAllDB()
       .then(() => successHandler(response, {
         message: resMsg.deleteAllSuccess,
